py_requirement.txt
# gcc_linux-64==7.5  # may need to ln -s in virtualenv/conda
# gxx_linux-64==7.5
# gast==0.2.2  # avoid batch-norm warning if any
numpy==1.21.5
six
wheel
pip
tqdm
pandas
psutil
pyyaml
matplotlib
scipy
scikit-learn<1
cython<3
setuptools<50
protobuf=3.7